refactor: extract branch resolution
